# netcore/src/OpenWealth/Controllers/CustomersApi.cs
/*
 * Custody Services
 *
 * This API is part of the OpenWealth APIs for the connectivity between custody banks and WealthTechs (e.g. Portfolio Management Systems). This API allows the user to receive data from custody banks regarding accounts and positions. The API is designed to be used for either update end of day data batches or single near-realtime account/position information. Furthermore, this API allows the user to receive data from custody banks regarding transactions. In this version (2.0) only transactions which include effective cash or securities movements are in scope and corporate actions without any cash impact are delivered only on a informative basis. 
 *
 * The version of the OpenAPI document: 3.0.0
 * Contact: openwealth@synpulse.com
 * Generated by: https://openapi-generator.tech
 */

using System;
using System.Collections.Generic;
using System.ComponentModel.DataAnnotations;
using Microsoft.AspNetCore.Authorization;
using Microsoft.AspNetCore.Mvc;
using Microsoft.AspNetCore.Http;
using System.Text.Json;
using OpenWealth.Attributes;
using OpenWealth.Models;

namespace OpenWealth.Controllers
{ 
    /// <summary>
    /// 
    /// </summary>
    [ApiController]
    public class CustomersApiController : ControllerBase
    { 
        /// <summary>
        /// Returns all customers (business partners) accessible for the logged in user
        /// </summary>
        /// <remarks>This endpoint returns a highlevel list of customers accessible for the logged in user. Paging is done based on the customer object, i.e. if limit is set to 1, then 1 customer will be returned per page. </remarks>
        /// <param name="xCorrelationID">Unique ID (defined by the caller) which will be reflected back in the response.</param>
        /// <param name="cursor">An opaque string value used for pagination.</param>
        /// <param name="limit">Maximum number of items to return.</param>
        /// <response code="200">Returns a list with all customers.</response>
        /// <response code="400">Bad Request - The server cannot or will not process the request due to something that is perceived to be a client error as malformed request syntax. </response>
        /// <response code="401">Unauthorized - The request has not been applied because it lacks valid authentication credentials for the target resource.</response>
        /// <response code="403">Forbidden - The server understood the request but refuses to authorize it.</response>
        /// <response code="404">Not Found - The origin server did not find a current representation for the target resource or is not willing to disclose that one exists. </response>
        /// <response code="405">Method Not Allowed - The method received in the request-line is known by the origin server but not supported by the target resource. </response>
        /// <response code="500">Internal Server Error - The server encountered an unexpected condition that prevented it from fulfilling the request.</response>
        /// <response code="501">Not Implemented - The server does not support the functionality required to fulfill the request.</response>
        /// <response code="503">Service Unavailable - The server is currently unable to handle the request due to a temporary overload or scheduled maintenance.</response>
        [HttpGet]
        [Route("/customers")]
        [ValidateModelState]
        [ProducesResponseType(statusCode: 200, type: typeof(List<Customer>))]
        [ProducesResponseType(statusCode: 400, type: typeof(CommonErrorResponse))]
        [ProducesResponseType(statusCode: 401, type: typeof(CommonErrorResponse))]
        [ProducesResponseType(statusCode: 403, type: typeof(CommonErrorResponse))]
        [ProducesResponseType(statusCode: 404, type: typeof(CommonErrorResponse))]
        [ProducesResponseType(statusCode: 405, type: typeof(CommonErrorResponse))]
        [ProducesResponseType(statusCode: 500, type: typeof(CommonErrorResponse))]
        [ProducesResponseType(statusCode: 501, type: typeof(CommonErrorResponse))]
        [ProducesResponseType(statusCode: 503, type: typeof(CommonErrorResponse))]
        public virtual IActionResult GetCustomers([FromHeader (Name = "X-Correlation-ID")][Required()]string xCorrelationID, [FromQuery (Name = "cursor")]string? cursor, [FromQuery (Name = "limit")]int? limit)
        {

            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default(List<Customer>));
            //TODO: Uncomment the next line to return response 400 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(400, default(CommonErrorResponse));
            //TODO: Uncomment the next line to return response 401 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(401, default(CommonErrorResponse));
            //TODO: Uncomment the next line to return response 403 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(403, default(CommonErrorResponse));
            //TODO: Uncomment the next line to return response 404 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(404, default(CommonErrorResponse));
            //TODO: Uncomment the next line to return response 405 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(405, default(CommonErrorResponse));
            //TODO: Uncomment the next line to return response 500 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(500, default(CommonErrorResponse));
            //TODO: Uncomment the next line to return response 501 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(501, default(CommonErrorResponse));
            //TODO: Uncomment the next line to return response 503 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(503, default(CommonErrorResponse));
            string exampleJson = null;
            exampleJson = "[ {\n  \"referenceCurrency\" : \"CHF\",\n  \"id\" : \"12345-6\"\n}, {\n  \"referenceCurrency\" : \"CHF\",\n  \"id\" : \"12345-6\"\n} ]";
            exampleJson = "Custom MIME type example not yet supported: application/problem+json";
            exampleJson = "Custom MIME type example not yet supported: application/problem+json";
            exampleJson = "Custom MIME type example not yet supported: application/problem+json";
            exampleJson = "Custom MIME type example not yet supported: application/problem+json";
            exampleJson = "Custom MIME type example not yet supported: application/problem+json";
            exampleJson = "Custom MIME type example not yet supported: application/problem+json";
            exampleJson = "Custom MIME type example not yet supported: application/problem+json";
            exampleJson = "Custom MIME type example not yet supported: application/problem+json";
            
            var example = exampleJson != null
            ? JsonSerializer.Deserialize<List<Customer>>(exampleJson)
            : default(List<Customer>);
            //TODO: Change the data returned
            return new ObjectResult(example);
        }

        /// <summary>
        /// Returns a specific customer accessible for the logged in user
        /// </summary>
        /// <remarks>This endpoint returns a single customer.</remarks>
        /// <param name="customerId">Customer ID Parameter.</param>
        /// <param name="xCorrelationID">Unique ID (defined by the caller) which will be reflected back in the response.</param>
        /// <response code="200">Returns a the specified customer.</response>
        /// <response code="400">Bad Request - The server cannot or will not process the request due to something that is perceived to be a client error as malformed request syntax. </response>
        /// <response code="401">Unauthorized - The request has not been applied because it lacks valid authentication credentials for the target resource.</response>
        /// <response code="403">Forbidden - The server understood the request but refuses to authorize it.</response>
        /// <response code="404">Not Found - The origin server did not find a current representation for the target resource or is not willing to disclose that one exists. </response>
        /// <response code="405">Method Not Allowed - The method received in the request-line is known by the origin server but not supported by the target resource. </response>
        /// <response code="500">Internal Server Error - The server encountered an unexpected condition that prevented it from fulfilling the request.</response>
        /// <response code="501">Not Implemented - The server does not support the functionality required to fulfill the request.</response>
        /// <response code="503">Service Unavailable - The server is currently unable to handle the request due to a temporary overload or scheduled maintenance.</response>
        [HttpGet]
        [Route("/customers/{customerId}")]
        [ValidateModelState]
        [ProducesResponseType(statusCode: 200, type: typeof(Customer))]
        [ProducesResponseType(statusCode: 400, type: typeof(CommonErrorResponse))]
        [ProducesResponseType(statusCode: 401, type: typeof(CommonErrorResponse))]
        [ProducesResponseType(statusCode: 403, type: typeof(CommonErrorResponse))]
        [ProducesResponseType(statusCode: 404, type: typeof(CommonErrorResponse))]
        [ProducesResponseType(statusCode: 405, type: typeof(CommonErrorResponse))]
        [ProducesResponseType(statusCode: 500, type: typeof(CommonErrorResponse))]
        [ProducesResponseType(statusCode: 501, type: typeof(CommonErrorResponse))]
        [ProducesResponseType(statusCode: 503, type: typeof(CommonErrorResponse))]
        public virtual IActionResult GetCustomersByCustomerId([FromRoute (Name = "customerId")][Required][StringLength(128, MinimumLength=1)]string customerId, [FromHeader (Name = "X-Correlation-ID")][Required()]string xCorrelationID)
        {

            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default(Customer));
            //TODO: Uncomment the next line to return response 400 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(400, default(CommonErrorResponse));
            //TODO: Uncomment the next line to return response 401 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(401, default(CommonErrorResponse));
            //TODO: Uncomment the next line to return response 403 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(403, default(CommonErrorResponse));
            //TODO: Uncomment the next line to return response 404 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(404, default(CommonErrorResponse));
            //TODO: Uncomment the next line to return response 405 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(405, default(CommonErrorResponse));
            //TODO: Uncomment the next line to return response 500 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(500, default(CommonErrorResponse));
            //TODO: Uncomment the next line to return response 501 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(501, default(CommonErrorResponse));
            //TODO: Uncomment the next line to return response 503 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(503, default(CommonErrorResponse));
            string exampleJson = null;
            exampleJson = "{\n  \"referenceCurrency\" : \"CHF\",\n  \"id\" : \"12345-6\"\n}";
            exampleJson = "Custom MIME type example not yet supported: application/problem+json";
            exampleJson = "Custom MIME type example not yet supported: application/problem+json";
            exampleJson = "Custom MIME type example not yet supported: application/problem+json";
            exampleJson = "Custom MIME type example not yet supported: application/problem+json";
            exampleJson = "Custom MIME type example not yet supported: application/problem+json";
            exampleJson = "Custom MIME type example not yet supported: application/problem+json";
            exampleJson = "Custom MIME type example not yet supported: application/problem+json";
            exampleJson = "Custom MIME type example not yet supported: application/problem+json";
            
            var example = exampleJson != null
            ? JsonSerializer.Deserialize<Customer>(exampleJson)
            : default(Customer);
            //TODO: Change the data returned
            return new ObjectResult(example);
        }
    }
}