# netcore/src/OpenWealth/wwwroot/openapi-original.json
{
  "openapi" : "3.0.0",
  "info" : {
    "contact" : {
      "email" : "openwealth@synpulse.com"
    },
    "description" : "This API is part of the OpenWealth APIs for the connectivity between custody banks and WealthTechs (e.g. Portfolio Management Systems). This API allows the user to receive data from custody banks regarding accounts and positions. The API is designed to be used for either update end of day data batches or single near-realtime account/position information. Furthermore, this API allows the user to receive data from custody banks regarding transactions. In this version (2.0) only transactions which include effective cash or securities movements are in scope and corporate actions without any cash impact are delivered only on a informative basis.\n",
    "license" : {
      "name" : "Apache 2.0",
      "url" : "https://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "title" : "Custody Services",
    "version" : "3.0.0"
  },
...
  "servers" : [ {
    "description" : "Your Server URL",
    "url" : "https://api.openwealth.ch"
  } ],
...
  "paths" : {
    "/customers" : {
      "get" : {
        "description" : "This endpoint returns a highlevel list of customers accessible for the logged in user. Paging is done based on the customer object, i.e. if limit is set to 1, then 1 customer will be returned per page.\n",
        "operationId" : "getCustomers",
        "parameters" : [ {
          "description" : "An opaque string value used for pagination.",
          "explode" : true,
          "in" : "query",
          "name" : "cursor",
          "required" : false,
          "schema" : {
            "type" : "string"
          },
          "style" : "form"
        }, {
          "description" : "Maximum number of items to return.",
          "explode" : true,
          "in" : "query",
          "name" : "limit",
          "required" : false,
          "schema" : {
            "format" : "int32",
            "minimum" : 1,
            "type" : "integer"
          },
          "style" : "form"
        }, {
          "description" : "Unique ID (defined by the caller) which will be reflected back in the response.",
          "explode" : false,
          "in" : "header",
          "name" : "X-Correlation-ID",
          "required" : true,
          "schema" : {
            "type" : "string"
          },
          "style" : "simple"
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/Customer"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "Returns a list with all customers.",
            "headers" : {
              "X-Correlation-ID" : {
                "$ref" : "#/components/headers/X-Correlation-Id"
              },
              "nextCursor" : {
                "$ref" : "#/components/headers/X-Next-Cursor"
              }
            }
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CommonErrorResponse"
                }
              }
            },
            "description" : "Bad Request - The server cannot or will not process the request due to something that is perceived to be a client error as malformed request syntax.\n",
            "headers" : {
              "Content-Type" : {
                "$ref" : "#/components/headers/Problem-Content-Type"
              },
              "Content-Language" : {
                "$ref" : "#/components/headers/Content-Language"
              },
              "X-Correlation-Id" : {
                "$ref" : "#/components/headers/X-Correlation-Id"
              }
            }
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CommonErrorResponse"
                }
              }
            },
            "description" : "Unauthorized - The request has not been applied because it lacks valid authentication credentials for the target resource.",
            "headers" : {
              "Content-Type" : {
                "$ref" : "#/components/headers/Problem-Content-Type"
              },
              "Content-Language" : {
                "$ref" : "#/components/headers/Content-Language"
              },
              "X-Correlation-Id" : {
                "$ref" : "#/components/headers/X-Correlation-Id"
              }
            }
          },
          "403" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CommonErrorResponse"
                }
              }
            },
            "description" : "Forbidden - The server understood the request but refuses to authorize it.",
            "headers" : {
              "Content-Type" : {
                "$ref" : "#/components/headers/Problem-Content-Type"
              },
              "Content-Language" : {
                "$ref" : "#/components/headers/Content-Language"
              },
              "X-Correlation-Id" : {
                "$ref" : "#/components/headers/X-Correlation-Id"
              }
            }
          },
          "404" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CommonErrorResponse"
                }
              }
            },
            "description" : "Not Found - The origin server did not find a current representation for the target resource or is not willing to disclose that one exists.\n",
            "headers" : {
              "Content-Type" : {
                "$ref" : "#/components/headers/Problem-Content-Type"
              },
              "Content-Language" : {
                "$ref" : "#/components/headers/Content-Language"
              },
              "X-Correlation-Id" : {
                "$ref" : "#/components/headers/X-Correlation-Id"
              }
            }
          },
          "405" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CommonErrorResponse"
                }
              }
            },
            "description" : "Method Not Allowed - The method received in the request-line is known by the origin server but not supported by the target resource.\n",
            "headers" : {
              "Content-Type" : {
                "$ref" : "#/components/headers/Problem-Content-Type"
              },
              "Content-Language" : {
                "$ref" : "#/components/headers/Content-Language"
              },
              "X-Correlation-Id" : {
                "$ref" : "#/components/headers/X-Correlation-Id"
              }
            }
          },
          "500" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CommonErrorResponse"
                }
              }
            },
            "description" : "Internal Server Error - The server encountered an unexpected condition that prevented it from fulfilling the request.",
            "headers" : {
              "Content-Type" : {
                "$ref" : "#/components/headers/Problem-Content-Type"
              },
              "Content-Language" : {
                "$ref" : "#/components/headers/Content-Language"
              },
              "X-Correlation-Id" : {
                "$ref" : "#/components/headers/X-Correlation-Id"
              }
            }
          },
          "501" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CommonErrorResponse"
                }
              }
            },
            "description" : "Not Implemented - The server does not support the functionality required to fulfill the request.",
            "headers" : {
              "Content-Type" : {
                "$ref" : "#/components/headers/Problem-Content-Type"
              },
              "Content-Language" : {
                "$ref" : "#/components/headers/Content-Language"
              },
              "X-Correlation-Id" : {
                "$ref" : "#/components/headers/X-Correlation-Id"
              }
            }
          },
          "503" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CommonErrorResponse"
                }
              }
            },
            "description" : "Service Unavailable - The server is currently unable to handle the request due to a temporary overload or scheduled maintenance.",
            "headers" : {
              "Content-Type" : {
                "$ref" : "#/components/headers/Problem-Content-Type"
              },
              "Content-Language" : {
                "$ref" : "#/components/headers/Content-Language"
              },
              "X-Correlation-Id" : {
                "$ref" : "#/components/headers/X-Correlation-Id"
              }
            }
          }
        },
        "summary" : "Returns all customers (business partners) accessible for the logged in user",
        "tags" : [ "customers" ]
      }
    },
    "/customers/{customerId}" : {
      "get" : {
        "description" : "This endpoint returns a single customer.",
        "operationId" : "getCustomersByCustomerId",
        "parameters" : [ {
          "description" : "Customer ID Parameter.",
          "explode" : false,
          "in" : "path",
          "name" : "customerId",
          "required" : true,
          "schema" : {
            "description" : "Unique and unambiguous identification used by the bank for the customer.",
            "example" : "123123-456",
            "maxLength" : 128,
            "minLength" : 1,
            "type" : "string"
          },
          "style" : "simple"
        }, {
          "description" : "Unique ID (defined by the caller) which will be reflected back in the response.",
          "explode" : false,
          "in" : "header",
          "name" : "X-Correlation-ID",
          "required" : true,
          "schema" : {
            "type" : "string"
          },
          "style" : "simple"
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Customer"
                }
              }
            },
            "description" : "Returns a the specified customer.",
            "headers" : {
              "X-Correlation-ID" : {
                "$ref" : "#/components/headers/X-Correlation-Id"
              }
            }
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CommonErrorResponse"
                }
              }
            },
            "description" : "Bad Request - The server cannot or will not process the request due to something that is perceived to be a client error as malformed request syntax.\n",
            "headers" : {
              "Content-Type" : {
                "$ref" : "#/components/headers/Problem-Content-Type"
              },
              "Content-Language" : {
                "$ref" : "#/components/headers/Content-Language"
              },
              "X-Correlation-Id" : {
                "$ref" : "#/components/headers/X-Correlation-Id"
              }
            }
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CommonErrorResponse"
                }
              }
            },
            "description" : "Unauthorized - The request has not been applied because it lacks valid authentication credentials for the target resource.",
            "headers" : {
              "Content-Type" : {
                "$ref" : "#/components/headers/Problem-Content-Type"
              },
              "Content-Language" : {
                "$ref" : "#/components/headers/Content-Language"
              },
              "X-Correlation-Id" : {
                "$ref" : "#/components/headers/X-Correlation-Id"
              }
            }
          },
          "403" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CommonErrorResponse"
                }
              }
            },
            "description" : "Forbidden - The server understood the request but refuses to authorize it.",
            "headers" : {
              "Content-Type" : {
                "$ref" : "#/components/headers/Problem-Content-Type"
              },
              "Content-Language" : {
                "$ref" : "#/components/headers/Content-Language"
              },
              "X-Correlation-Id" : {
                "$ref" : "#/components/headers/X-Correlation-Id"
              }
            }
          },
          "404" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CommonErrorResponse"
                }
              }
            },
            "description" : "Not Found - The origin server did not find a current representation for the target resource or is not willing to disclose that one exists.\n",
            "headers" : {
              "Content-Type" : {
                "$ref" : "#/components/headers/Problem-Content-Type"
              },
              "Content-Language" : {
                "$ref" : "#/components/headers/Content-Language"
              },
              "X-Correlation-Id" : {
                "$ref" : "#/components/headers/X-Correlation-Id"
              }
            }
          },
          "405" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CommonErrorResponse"
                }
              }
            },
            "description" : "Method Not Allowed - The method received in the request-line is known by the origin server but not supported by the target resource.\n",
            "headers" : {
              "Content-Type" : {
                "$ref" : "#/components/headers/Problem-Content-Type"
              },
              "Content-Language" : {
                "$ref" : "#/components/headers/Content-Language"
              },
              "X-Correlation-Id" : {
                "$ref" : "#/components/headers/X-Correlation-Id"
              }
            }
          },
          "500" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CommonErrorResponse"
                }
              }
            },
            "description" : "Internal Server Error - The server encountered an unexpected condition that prevented it from fulfilling the request.",
            "headers" : {
              "Content-Type" : {
                "$ref" : "#/components/headers/Problem-Content-Type"
              },
              "Content-Language" : {
                "$ref" : "#/components/headers/Content-Language"
              },
              "X-Correlation-Id" : {
                "$ref" : "#/components/headers/X-Correlation-Id"
              }
            }
          },
          "501" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CommonErrorResponse"
                }
              }
            },
            "description" : "Not Implemented - The server does not support the functionality required to fulfill the request.",
            "headers" : {
              "Content-Type" : {
                "$ref" : "#/components/headers/Problem-Content-Type"
              },
              "Content-Language" : {
                "$ref" : "#/components/headers/Content-Language"
              },
              "X-Correlation-Id" : {
                "$ref" : "#/components/headers/X-Correlation-Id"
              }
            }
          },
          "503" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CommonErrorResponse"
                }
              }
            },
            "description" : "Service Unavailable - The server is currently unable to handle the request due to a temporary overload or scheduled maintenance.",
            "headers" : {
              "Content-Type" : {
                "$ref" : "#/components/headers/Problem-Content-Type"
              },
              "Content-Language" : {
                "$ref" : "#/components/headers/Content-Language"
              },
              "X-Correlation-Id" : {
                "$ref" : "#/components/headers/X-Correlation-Id"
              }
            }
          }
        },
        "summary" : "Returns a specific customer accessible for the logged in user",
        "tags" : [ "customers" ]
      }
    },
...
  },
  "components" : {
    "headers" : {
      "X-Correlation-Id" : {
        "description" : "Client defined ID from request to correlates HTTP requests between a client and server.",
        "explode" : false,
        "schema" : {
          "example" : "f058ebd6-02f7-4d3f-942e-904344e8cde5",
          "type" : "string"
        },
        "style" : "simple"
      },
      "X-Next-Cursor" : {
        "description" : "An opaque string value, or an empty string if there are no more results.",
        "explode" : false,
        "required" : false,
        "schema" : {
          "type" : "string"
        },
        "style" : "simple"
      },
      "Problem-Content-Type" : {
        "description" : "application/problem+json; charset=utf-8 according to RFC7807.",
        "explode" : false,
        "schema" : {
          "example" : "application/problem+json",
          "type" : "string"
        },
        "style" : "simple"
      },
      "Content-Language" : {
        "description" : "Response language - always en.",
        "explode" : false,
        "schema" : {
          "example" : "en",
          "type" : "string"
        },
        "style" : "simple"
      }
    },
...
    "schemas" : {
...
      "Customer" : {
        "description" : "Overview of the customer with the respective accounts.",
        "example" : {
          "referenceCurrency" : "CHF",
          "id" : "12345-6"
        },
        "properties" : {
          "id" : {
            "description" : "Unique and unambiguous identification used by the bank for the customer.",
            "example" : "12345-6",
            "maxLength" : 128,
            "type" : "string"
          },
          "referenceCurrency" : {
            "description" : "ISO 4217 code.",
            "example" : "CHF",
            "pattern" : "^[A-Z]{3}$",
            "title" : "Currency",
            "type" : "string"
          }
        },
        "required" : [ "id" ],
        "title" : "Customer",
        "type" : "object"
      },
      "CommonErrorType" : {
        "description" : "Error Types for CommonErrorResponse.",
        "enum" : [ "/problems/INVALID_PAYLOAD", "/problems/MALFORMED_PAYLOAD", "/problems/INVALID_TOKEN", "/problems/EXPIRED_TOKEN", "/problems/INSUFFICIENT_PRIVILEGES", "/problems/NO_ACCESS_TO_RESOURCE", "/problems/RESOURCE_DOES_NOT_EXIST", "/problems/RESOURCE_NOT_READY", "/problems/RESOURCE_TOO_LARGE", "/problems/WRONG_METHOD", "/problems/OPERATION_NOT_ALLOWED", "/problems/TECHNICAL_ERROR", "/problems/NOT_IMPLEMENTED", "/problems/SERVICE_UNAVAILABLE" ],
        "example" : "/problems/TECHNICAL_ERROR",
        "title" : "Common Error Type",
        "type" : "string"
      },
      "CommonErrorResponse" : {
        "description" : "Common error response.",
        "properties" : {
          "type" : {
            "$ref" : "#/components/schemas/CommonErrorType"
          },
          "title" : {
            "description" : "Title of the error response.",
            "example" : "This is the general problem description",
            "type" : "string"
          },
          "detail" : {
            "description" : "Detailed description of the error response.",
            "example" : "Detailed problem description with respect to the current request",
            "type" : "string"
          },
          "instance" : {
            "description" : "Entity instance that threw the error.",
            "example" : "path/to/corresponding/resource",
            "type" : "string"
          }
        },
        "title" : "CommonErrorResponse",
        "type" : "object"
      },
...
    },
    "securitySchemes" : { }
  }
}